First working version of Sigfox uplink DBPSK modulation.

+++ inc/peripherals/lptim.h
@@ -15,6 +15,6 @@ void LPTIM1_Start(unsigned short period_us);
 unsigned char LPTIM1_GetArrmFlag(void);
 void LPTIM1_ClearArrmFlag(void);
 void LPTIM1_Stop(void);
-unsigned short LPTIM1_GetMicroseconds(void);
+unsigned int LPTIM1_GetMicroseconds(void);
 
 #endif /* LPTIM_H */
